+++ url_normalize/normalize_path.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from .tools import quote, unquote
+from .tools import RESERVED_CHARACTERS, normalize_component
 
 
 def normalize_path(path: str, scheme: str) -> str:
@@ -21,7 +21,7 @@ def normalize_path(path: str, scheme: str) -> str:
     # Only perform percent-encoding where it is essential.
     # Always use uppercase A-through-F characters when percent-encoding.
     # All portions of the URI must be utf-8 encoded NFC from Unicode strings
-    path = quote(unquote(path), "~:/#[]@!$&'()*+,;=")
+    path = normalize_component(path, "~:/[]@!$&'()*+,;=", RESERVED_CHARACTERS)
     # Prevent dot-segments appearing in non-relative URI paths.
     if scheme in {"", "http", "https", "ftp", "file"}:
         output: list[str] = []